requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asgiref==3.8.1
astroid==3.0.3
attrs==25.3.0
autopep8==2.0.4
beautifulsoup4==4.13.4
billiard==4.2.1
Brotli==1.1.0
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.2
cssselect2==0.8.0
defusedxml==0.7.1
diff-match-patch==20241021
dill==0.3.8
distlib==0.3.9
dj-database-url==3.0.0
Django==5.2.1
django-apscheduler==0.7.0
django-cors-headers==4.7.0
django-extensions==3.2.3
django-heroku==0.3.1
django-import-export==4.2.0
django-taggit==6.1.0
django-unfold==0.63.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
et_xmlfile==2.0.0
Faker==37.3.0
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
geoip2==5.1.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isort==5.13.2
kombu==5.5.3
maxminddb==2.7.0
mccabe==0.7.0
multidict==6.4.4
numpy==2.3.2
oauthlib==3.2.2
ollama==0.5.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pep8==1.7.1
pillow==11.2.1
platformdirs==4.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary>=2.9.9
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
PyJWT==2.9.0
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyphen==0.17.2
PySocks==1.7.1
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
qrcode==8.2
requests==2.32.3
requests-oauthlib==2.0.0
selenium==4.32.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
tablib==3.7.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tomlkit==0.12.3
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
vine==5.1.0
virtualenv==20.30.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
websocket-client==1.8.0
whitenoise==6.9.0
wsproto==1.2.0
xlrd==2.0.2
xlsxwriter==3.2.5
xlwt==1.3.0
yarl==1.20.1
zopfli==0.2.3.post1
wfastcgi==3.0.0